# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Correlator API
  version: 0.1.0
  description: |
    **Correlator Incident Correlation Engine API**

    Correlator automatically connects test failures to job runs and downstream impact
    across your data stack (dbt, Airflow, Great Expectations).

    ## Authentication

    Protected endpoints require API key authentication via `Authorization: Bearer` header.
    Public health endpoints (`/ping`, `/ready`, `/health`) do not require authentication.

    ## Base URLs

    - API endpoints: `/api/v1/*`
    - Health probes: `/ping`, `/ready`, `/health`
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/lineage:
    post:
      summary: Ingest single OpenLineage event
      description: |
        Standard OpenLineage single-event ingestion endpoint. Accepts a single
        OpenLineage RunEvent JSON object.

        This endpoint is compatible with standard OpenLineage clients configured with:
        ```
        OPENLINEAGE_URL=http://correlator:8080
        ```

        Events are processed with idempotency - duplicate events return 200 OK.

        **Request Limits:**
        - Max request body: 1 MB
      operationId: ingestLineageEvent
      tags:
        - OpenLineage Ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineageEvent'
            example:
              eventTime: "2024-01-01T12:00:00Z"
              producer: "https://github.com/OpenLineage/OpenLineage/tree/1.0.0/integration/dbt"
              schemaURL: "https://openlineage.io/spec/2-0-2/OpenLineage.json"
              eventType: "COMPLETE"
              run:
                runId: "550e8400-e29b-41d4-a716-446655440000"
              job:
                namespace: "dbt_prod"
                name: "models.users"
              inputs:
                - namespace: "postgresql://prod-db/public"
                  name: "raw_users"
              outputs:
                - namespace: "postgresql://prod-db/public"
                  name: "clean_users"
      responses:
        '200':
          description: Event processed successfully (empty body per OL spec)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          description: Event failed validation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/lineage/batch:
    post:
      summary: Ingest batch OpenLineage events
      description: |
        Batch event ingestion endpoint. Accepts a JSON array of OpenLineage RunEvents.

        Events are processed with idempotency - duplicate events are detected and counted
        as successful (OpenLineage specification behavior).

        **Request Limits:**
        - Max batch size: 1000 events
        - Max request body: 1 MB
      operationId: ingestLineageEventBatch
      tags:
        - OpenLineage Ingestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LineageEvent'
              minItems: 1
              maxItems: 1000
            example:
              - eventTime: "2024-01-01T12:00:00Z"
                producer: "https://github.com/OpenLineage/OpenLineage/tree/1.0.0/integration/dbt"
                schemaURL: "https://openlineage.io/spec/2-0-2/OpenLineage.json"
                eventType: "COMPLETE"
                run:
                  runId: "550e8400-e29b-41d4-a716-446655440000"
                job:
                  namespace: "dbt_prod"
                  name: "models.users"
                inputs:
                  - namespace: "postgresql://prod-db/public"
                    name: "raw_users"
                outputs:
                  - namespace: "postgresql://prod-db/public"
                    name: "clean_users"
      responses:
        '200':
          description: All events processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageResponse'
        '207':
          description: Partial success - some events failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/incidents:
    get:
      summary: List incidents
      description: |
        Retrieve paginated list of incidents with optional filtering by status, time range,
        and time window.

        The view only returns failed/error test results that have lineage correlation.

        **Retry Deduplication:** When a test runs multiple times under the same orchestrator
        run (e.g., Airflow retries), only the latest attempt is returned. The `retry_context`
        field provides metadata about the retry group.

        **Status Filtering:**
        - `active` (default): Open + acknowledged incidents
        - `resolved`: Resolved incidents (within `window` days)
        - `muted`: Muted incidents (within `window` days)
        - `all`: All statuses (active always shown; resolved/muted within `window` days)
      operationId: listIncidents
      tags:
        - Correlation Queries
      parameters:
        - name: status
          in: query
          description: Filter by resolution status
          schema:
            type: string
            enum: [active, resolved, muted, all]
            default: active
        - name: window
          in: query
          description: |
            Time window in days for historical views (resolved/muted). Defaults to 7 for
            non-active filters. Ignored when status=active.
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
        - name: since
          in: query
          description: Filter incidents after this timestamp (ISO8601)
          schema:
            type: string
            format: date-time
          example: "2024-01-01T00:00:00Z"
        - name: limit
          in: query
          description: Maximum number of incidents to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of incidents to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of incidents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/incidents/{id}:
    get:
      summary: Get incident details
      description: |
        Retrieve detailed incident information including test details, dataset info,
        producing job, upstream/downstream lineage, resolution status, and retry context.

        The `correlation_status` field indicates:
        - `correlated`: Fully correlated with a data-producing job
        - `orphan`: Namespace aliasing issue (needs configuration)
        - `unknown`: Correlation status cannot be determined

        The `retry_context` field (if present) includes `other_attempts` — an array of
        sibling test executions under the same orchestrator run.
      operationId: getIncidentDetails
      tags:
        - Correlation Queries
      parameters:
        - name: id
          in: path
          required: true
          description: Test result ID (numeric)
          schema:
            type: string
          example: "123"
      responses:
        '200':
          description: Incident details with lineage and resolution info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/incidents/{id}/status:
    patch:
      summary: Update incident resolution status
      description: |
        Transition an incident to a new resolution status.

        **Valid transitions:**
        - `open` → `acknowledged`, `resolved`, `muted`
        - `acknowledged` → `resolved`, `muted`
        - `resolved` and `muted` are terminal (no further transitions in alpha)

        **Status-specific rules:**
        - `acknowledged`: No additional fields required
        - `resolved`: `reason` optional (defaults to `"manual"`)
        - `muted`: `reason` required (e.g., `"false_positive"`, `"expected"`),
          `mute_days` optional (defaults to 30, max 365)

        Returns `409 Conflict` if the transition is not allowed from the current state.
      operationId: updateIncidentStatus
      tags:
        - Incident Resolution
      parameters:
        - name: id
          in: path
          required: true
          description: Test result ID (numeric)
          schema:
            type: string
          example: "123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
            examples:
              acknowledge:
                summary: Acknowledge an incident
                value:
                  status: acknowledged
              resolve:
                summary: Resolve an incident
                value:
                  status: resolved
                  reason: manual
                  note: "Fixed upstream data source"
              mute:
                summary: Mute an incident
                value:
                  status: muted
                  reason: false_positive
                  mute_days: 7
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/incidents/counts:
    get:
      summary: Get incident counts by status
      description: |
        Returns aggregate counts of incidents grouped by resolution status.

        Active incidents (open + acknowledged) are always counted regardless of time window.
        Resolved and muted counts are scoped to the last 30 days.
      operationId: getIncidentCounts
      tags:
        - Incident Resolution
      responses:
        '200':
          description: Incident counts by status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentCountsResponse'
              example:
                active: 12
                resolved: 45
                muted: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication via `Authorization: Bearer <key>` header.
        Compatible with standard OpenLineage clients configured with OPENLINEAGE_API_KEY.
  schemas:
...
    LineageEvent:
      type: object
      required:
        - eventTime
        - eventType
        - run
        - job
      properties:
        eventTime:
          type: string
          format: date-time
          description: Event timestamp (ISO8601)
        eventType:
          type: string
          enum: [START, RUNNING, COMPLETE, ABORT, FAIL, OTHER]
          description: Event type
        producer:
          type: string
          description: Producer URL identifying the emitting system
        schemaURL:
          type: string
          description: OpenLineage schema URL
        run:
          $ref: '#/components/schemas/Run'
        job:
          $ref: '#/components/schemas/Job'
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/Dataset'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/Dataset'
    Run:
      type: object
      required:
        - runId
      properties:
        runId:
          type: string
          description: Unique run identifier
        facets:
          type: object
          additionalProperties: true
          description: Run facets (OpenLineage spec)
    Job:
      type: object
      required:
        - namespace
        - name
      properties:
        namespace:
          type: string
          description: Job namespace
        name:
          type: string
          description: Job name
        facets:
          type: object
          additionalProperties: true
          description: Job facets (OpenLineage spec)
    Dataset:
      type: object
      required:
        - namespace
        - name
      properties:
        namespace:
          type: string
          description: Dataset namespace (e.g., postgresql://prod-db/public)
        name:
          type: string
          description: Dataset name (e.g., users)
        facets:
          type: object
          additionalProperties: true
        inputFacets:
          type: object
          additionalProperties: true
        outputFacets:
          type: object
          additionalProperties: true
    LineageResponse:
      type: object
      required:
        - status
        - summary
        - failed_events
        - correlation_id
        - timestamp
      properties:
        status:
          type: string
          enum: [success, error]
          description: Overall request status
        summary:
          $ref: '#/components/schemas/ResponseSummary'
        failed_events:
          type: array
          items:
            $ref: '#/components/schemas/FailedEvent'
        correlation_id:
          type: string
          description: Request correlation ID for tracing
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ResponseSummary:
      type: object
      required:
        - received
        - successful
        - failed
        - retriable
        - non_retriable
      properties:
        received:
          type: integer
          description: Total events received
        successful:
          type: integer
          description: Successfully processed events
        failed:
          type: integer
          description: Failed events
        retriable:
          type: integer
          description: Retriable failures
        non_retriable:
          type: integer
          description: Non-retriable failures (validation errors)
    FailedEvent:
      type: object
      required:
        - index
        - reason
        - retriable
      properties:
        index:
          type: integer
          description: Zero-based index of failed event
        reason:
          type: string
          description: Failure reason
        retriable:
          type: boolean
          description: Whether failure is retriable

    # Incident Schemas
    IncidentListResponse:
      type: object
      required:
        - incidents
        - total
        - limit
        - offset
        - orphan_count
      properties:
        incidents:
          type: array
          items:
            $ref: '#/components/schemas/IncidentSummary'
        total:
          type: integer
          description: Total number of incidents matching filter
        limit:
          type: integer
          description: Page size
        offset:
          type: integer
          description: Page offset
        orphan_count:
          type: integer
          description: Number of datasets with test failures but no producer correlation (not shown in incidents list)
...
    IncidentDetailResponse:
      type: object
      required:
        - id
        - test
        - dataset
        - upstream
        - downstream
        - correlation_status
        - resolution_status
      properties:
        id:
          type: string
          description: Test result ID
        test:
          $ref: '#/components/schemas/TestDetail'
        dataset:
          $ref: '#/components/schemas/DatasetDetail'
        job:
          $ref: '#/components/schemas/JobDetail'
          nullable: true
          description: Null if uncorrelated
        upstream:
          type: array
          description: Datasets consumed to produce the tested dataset (data provenance)
          items:
            $ref: '#/components/schemas/UpstreamDataset'
        downstream:
          type: array
          description: Datasets affected by this dataset (impact analysis)
          items:
            $ref: '#/components/schemas/DownstreamDataset'
        correlation_status:
          type: string
          enum: [correlated, orphan, unknown]
          description: |
            - correlated: Fully correlated with data-producing job
            - orphan: Namespace aliasing issue
            - unknown: Cannot determine correlation
        resolution_status:
          type: string
          enum: [open, acknowledged, resolved, muted]
          description: Current resolution status
        resolved_by:
          type: string
          description: Who resolved ("auto" or client ID). Omitted if open.
        resolution_reason:
          type: string
          description: Resolution reason (e.g., "manual", "auto_pass", "false_positive"). Omitted if open.
        resolved_at:
          type: string
          format: date-time
          nullable: true
          description: When the resolution status was last changed
        mute_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Mute expiry timestamp (only for muted status)
        retry_context:
          $ref: '#/components/schemas/RunRetryContextDetail'
          nullable: true
          description: |
            Full retry metadata including other attempts. Null when no retries exist.
    TestDetail:
      type: object
      required:
        - name
        - type
        - status
        - message
        - executed_at
        - duration_ms
        - producer
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        message:
          type: string
        executed_at:
          type: string
          format: date-time
        duration_ms:
          type: integer
          format: int64
        producer:
          type: string
          description: Tool that ran this test (e.g., "great_expectations", "dbt")
    DatasetDetail:
      type: object
      required:
        - urn
        - name
        - namespace
      properties:
        urn:
          type: string
        name:
          type: string
        namespace:
          type: string
...
    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [acknowledged, resolved, muted]
          description: Target resolution status
        reason:
          type: string
          description: |
            Resolution reason. Required for `muted` (e.g., "false_positive", "expected").
            Optional for `resolved` (defaults to "manual").
        note:
          type: string
          description: Free-text note about the resolution
        mute_days:
          type: integer
          minimum: 1
          maximum: 365
          default: 30
          description: Mute duration in days (only for muted status)
    UpdateStatusResponse:
      type: object
      required:
        - id
        - resolution_status
        - resolved_by
        - resolved_at
      properties:
        id:
          type: string
          description: Test result ID
        resolution_status:
          type: string
          enum: [acknowledged, resolved, muted]
          description: New resolution status
        resolved_by:
          type: string
          description: Who made the change (client ID from API key)
        resolved_at:
          type: string
          format: date-time
          description: Timestamp of the status change
        resolution_reason:
          type: string
          nullable: true
          description: Resolution reason (null for acknowledged)
        mute_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Mute expiry timestamp (only for muted status)
    IncidentCountsResponse:
      type: object
      required:
        - active
        - resolved
        - muted
      properties:
        active:
          type: integer
          description: Count of open + acknowledged incidents
        resolved:
          type: integer
          description: Count of resolved incidents (within 30-day window)
        muted:
          type: integer
          description: Count of muted incidents (within 30-day window)

    # Run Retry Schemas
...
    Error:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: Problem type URI
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Detailed explanation
        instance:
          type: string
          description: URI of the specific occurrence
        correlation_id:
          type: string
          description: Request correlation ID
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/400"
            title: "Bad Request"
            status: 400
            detail: "Invalid parameter 'limit': must be between 1 and 100"
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/401"
            title: "Unauthorized"
            status: 401
            detail: "Missing API key"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/404"
            title: "Not Found"
            status: 404
            detail: "Incident not found"
    Conflict:
      description: State transition not allowed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/409"
            title: "Conflict"
            status: 409
            detail: "cannot transition from resolved to acknowledged: resolved is a terminal state"
    UnprocessableEntity:
      description: Validation error in request body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/422"
            title: "Unprocessable Entity"
            status: 422
            detail: "status must be one of: acknowledged, resolved, muted"
    PayloadTooLarge:
      description: Request body exceeds 1 MB limit
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/413"
            title: "Payload Too Large"
            status: 413
            detail: "Request body exceeds maximum size of 1048576 bytes"
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/415"
            title: "Unsupported Media Type"
            status: 415
            detail: "Content-Type must be application/json"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/429"
            title: "Rate Limited"
            status: 429
            detail: "Rate limit exceeded"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "https://getcorrelator.io/problems/500"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred"